// state/Slider/slider.action-types.ts
export enum ActionTypes {
    SLIDER_UPLOAD_REQUEST = 'SLIDER_UPLOAD_REQUEST',
    SLIDER_UPLOAD_SUCCESS = 'SLIDER_UPLOAD_SUCCESS',
    SLIDER_UPLOAD_FAIL = 'SLIDER_UPLOAD_FAIL',

    SLIDER_FETCH_REQUEST = 'SLIDER_FETCH_REQUEST',
    SLIDER_FETCH_SUCCESS = 'SLIDER_FETCH_SUCCESS',
    SLIDER_FETCH_FAIL = 'SLIDER_FETCH_FAIL',

    SLIDER_DELETE_REQUEST = 'Slider_DELETE_REQUEST',
    SLIDER_DELETE_SUCCESS = 'Slider_DELETE_SUCCESS',
    SLIDER_DELETE_FAIL = 'Slider_DELETE_FAIL',

    SLIDER_RESET = 'SLIDER_RESET',
  }
  